refactor(lib/oauth): move protocol to lib; brick re-exports
